fix: sortTimePlans 를 모든 플랜 정렬에 사용할 수 있도록 sortPlans로 변경 (#39)
